+++ changelog.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+
+
+## [0.6.6] - 2018-04-18
+### Added
+- physical keyboard mapping
+- refactor input handling to it's own file
+- css color variables all in one place
+- try/catch for when a chord is not found + ui clear
+### Changed
+- redo keyboard ui
+
 ## [0.6.5] - 2018-03-21
 ### Added
 - separate css for white key sizing
@@ -13,11 +25,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Deleted
 - deleted keyboard ui numbers
 
+
 ## [0.6.4] - 2018-03-03
 ### Added
 - all method tests (4 passing!)
 ### Deleted
 - deleted `getUserChord()`; no longer used
+
 
 ## [0.6.3] - 2018-03-01
 ### Added
@@ -25,12 +39,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - first unit test passed!
 
+
 ## [0.6.2] - 2018-02-28
 ### Added
 - node to start unit testing
 - jest unit tests; setup only
 ### Changed
 - global variables now in one file, `globals.js`
+
 
 ## [0.6.1] - 2018-02-27
 ### Added
@@ -101,6 +117,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Commits exist before 0.5.5 but I wasn't keeping count.
 
+[0.6.6]: #
 [0.6.5]: #
 [0.6.4]: https://github.com/ManuelVargas1251/Chord-Finder/commit/656dc07b98d596fb1f9e3a93e5081324f08fc9ef
 [0.6.3]: https://github.com/ManuelVargas1251/Chord-Finder/commit/21e4a01ccd3bb22fd34d4a1f3c75feaec14a85e9